Refactor to remove duplicate memory allocator
