+++ tests/conftest.py
@@ -13,6 +13,7 @@ import pytest
 
 # First Party
 from lmcache.v1.cache_engine import LMCacheEngineBuilder
+from lmcache.v1.memory_management import MixedMemoryAllocator
 
 # This is to mock the constructor and destructor of
 # MixedMemoryAllocator and PinMemoryAllocator to
@@ -359,3 +360,27 @@ def autorelease_v1(request):
     # Cleanup all objects created by the factory
     # for obj in objects:
     #    obj.close()
+
+
+@pytest.fixture(scope="session")
+def memory_allocator():
+    """One MixedMemoryAllocator (1GB) for the whole test session;
+    .close() is a no-op per-test."""
+    _real = MixedMemoryAllocator(1024 * 1024 * 1024)  # 1GB
+
+    class _NoCloseWrapper:
+        def __init__(self, real):
+            self._real = real
+
+        def __getattr__(self, name):
+            return getattr(self._real, name)
+
+        def close(self):
+            # No-op so per-test close() calls don't shut down the shared allocator
+            pass
+
+    try:
+        yield _NoCloseWrapper(_real)
+    finally:
+        # Actually close once when the session ends
+        _real.close()
