+++ tests/v1/storage_backend/test_local_cpu_backend.py
@@ -14,7 +14,6 @@ from lmcache.v1.memory_management import (
     AdHocMemoryAllocator,
     MemoryFormat,
     MemoryObj,
-    MixedMemoryAllocator,
 )
 from lmcache.v1.storage_backend.local_cpu_backend import LocalCPUBackend
 
@@ -63,30 +62,6 @@ def create_test_memory_obj(shape=(2, 16, 8, 128), dtype=torch.bfloat16) -> Memor
     allocator = AdHocMemoryAllocator(device="cpu")
     memory_obj = allocator.allocate(shape, dtype, fmt=MemoryFormat.KV_T2D)
     return memory_obj
-
-
-@pytest.fixture(scope="session")
-def memory_allocator():
-    """One MixedMemoryAllocator (1GB) for the whole test session;
-    .close() is a no-op per-test."""
-    _real = MixedMemoryAllocator(1024 * 1024 * 1024)  # 1GB
-
-    class _NoCloseWrapper:
-        def __init__(self, real):
-            self._real = real
-
-        def __getattr__(self, name):
-            return getattr(self._real, name)
-
-        def close(self):
-            # No-op so per-test close() calls don't shut down the shared allocator
-            pass
-
-    try:
-        yield _NoCloseWrapper(_real)
-    finally:
-        # Actually close once when the session ends
-        _real.close()
 
 
 @pytest.fixture
